backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "coverage>=7.13.2",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = [
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-fail-under=60",
    "--cov-config=pyproject.toml",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*tests.py",
    "conftest.py",
    "manage.py",
    "core/wsgi.py",
    "core/asgi.py",
]

[tool.coverage.report]
fail_under = 60
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
